requirements.txt
# Copyright 2021 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

ansible==2.9.11
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
google-auth==1.26.0  
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.11
google-cloud-secret-manager==2.11.1
